pyproject.toml
[project]
name = "mise-en-prod-3a"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "duckdb>=1.4.4",
    "joblib>=1.5.3",
    "numpy>=2.0.2",
    "pandas>=2.3.3",
    "pyarrow>=21.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "s3fs>=2025.10.0",
    "scikit-learn>=1.6.1",
    "xgboost>=2.1.4",
]
